实现订单提交与查看

+++ backend/src/main/java/com/example/commerce/model/CartItem.java
@@ -1,5 +1,7 @@
 package com.example.commerce.model;
+
 import lombok.Data;
+
 import java.time.Instant;
 
 @Data
@@ -7,6 +9,6 @@ public class CartItem {
     private Long id;
     private Long userId;
     private Long productVariantId;
-    private Long quantity;
+    private Integer quantity;
     private Instant addedAt;
 }
